Address review suggestions: Move QJITDevice to pennylane_extensions
